Just split LTN save/load UI code into modules
