+++ apps/ltn/src/save/mod.rs
@@ -1,4 +1,6 @@
 mod perma;
+mod proposals_ui;
+mod save_dialog;
 mod share;
 
 use std::collections::BTreeSet;
@@ -8,13 +10,9 @@ use serde::{Deserialize, Serialize};
 
 use abstio::MapName;
 use abstutil::{Counter, Timer};
-use map_gui::tools::{FilePicker, FileSaver, FileSaverContents};
 use map_model::{BuildingID, EditRoad, Map};
-use widgetry::tools::{ChooseSomething, PopupMsg};
-use widgetry::{
-    lctrl, Choice, DrawBaselayer, EventCtx, GfxCtx, Key, Line, MultiKey, Outcome, Panel, State,
-    TextBox, Widget,
-};
+use widgetry::tools::PopupMsg;
+use widgetry::{EventCtx, State};
 
 use crate::logic::{BlockID, Partitioning};
 use crate::{pages, App, Edits, Transition};
@@ -149,219 +147,6 @@ fn stash_current_proposal(app: &mut App) {
         .unwrap() = Some(app.per_map.proposals.current_proposal.clone());
 }
 
-fn switch_to_existing_proposal(ctx: &mut EventCtx, app: &mut App, idx: usize) {
-    stash_current_proposal(app);
-
-    let proposal = app
-        .per_map
-        .proposals
-        .list
-        .get_mut(idx)
-        .unwrap()
-        .take()
-        .unwrap();
-    app.per_map.proposals.current = idx;
-
-    proposal.make_active(ctx, app);
-}
-
-struct SaveDialog {
-    panel: Panel,
-    preserve_state: PreserveState,
-    can_overwrite: bool,
-}
-
-impl SaveDialog {
-    fn new_state(
-        ctx: &mut EventCtx,
-        app: &App,
-        preserve_state: PreserveState,
-    ) -> Box<dyn State<App>> {
-        let parent = app
-            .per_map
-            .proposals
-            .current_proposal
-            .unsaved_parent
-            .clone();
-        let can_overwrite = parent.is_some() && parent != Some("existing LTNs".to_string());
-
-        let mut state = Self {
-            panel: Panel::new_builder(Widget::col(vec![
-                Widget::row(vec![
-                    Line("Save proposal").small_heading().into_widget(ctx),
-                    ctx.style().btn_close_widget(ctx),
-                ]),
-                if can_overwrite {
-                    Widget::row(vec![
-                        ctx.style()
-                            .btn_solid_destructive
-                            .text(format!("Overwrite \"{}\"", parent.unwrap()))
-                            .build_widget(ctx, "Overwrite"),
-                        Line("Or save a new copy below")
-                            .secondary()
-                            .into_widget(ctx)
-                            .centered_vert(),
-                    ])
-                } else {
-                    Widget::nothing()
-                },
-                Widget::row(vec![
-                    TextBox::default_widget(ctx, "input", String::new()),
-                    Widget::placeholder(ctx, "Save as"),
-                ]),
-                Widget::placeholder(ctx, "warning"),
-            ]))
-            .build(ctx),
-            preserve_state,
-            can_overwrite,
-        };
-        state.name_updated(ctx);
-        Box::new(state)
-    }
-
-    fn name_updated(&mut self, ctx: &mut EventCtx) {
-        let name = self.panel.text_box("input");
-
-        let warning = if name == "existing LTNs" {
-            Some("You can't overwrite the name \"existing LTNs\"")
-        } else if name.is_empty() {
-            Some("You have to name this proposal")
-        } else {
-            None
-        };
-
-        let btn = ctx
-            .style()
-            .btn_solid_primary
-            .text("Save as")
-            .disabled(warning.is_some())
-            .hotkey(if self.can_overwrite {
-                None
-            } else {
-                Some(MultiKey::from(Key::Enter))
-            })
-            .build_def(ctx);
-        self.panel.replace(ctx, "Save as", btn);
-
-        if let Some(warning) = warning {
-            self.panel
-                .replace(ctx, "warning", Line(warning).into_widget(ctx));
-        } else {
-            self.panel
-                .replace(ctx, "warning", Widget::placeholder(ctx, "warning"));
-        }
-    }
-
-    fn error(&self, ctx: &mut EventCtx, app: &mut App, err: impl AsRef<str>) -> Transition {
-        Transition::Multi(vec![
-            self.preserve_state.switch_to_state(ctx, app),
-            Transition::Push(PopupMsg::new_state(ctx, "Error", vec![err])),
-        ])
-    }
-}
-
-impl State<App> for SaveDialog {
-    fn event(&mut self, ctx: &mut EventCtx, app: &mut App) -> Transition {
-        match self.panel.event(ctx) {
-            Outcome::Clicked(x) => match x.as_ref() {
-                "close" => Transition::Pop,
-                "Save as" => {
-                    let name = self.panel.text_box("input");
-
-                    // TODO If we're clobbering something that exists in Proposals especially...
-                    // watch out
-
-                    app.per_map.proposals.current_proposal.name = name;
-                    app.per_map.proposals.current_proposal.unsaved_parent = None;
-                    return match inner_save(app) {
-                        // If we changed the name, we'll want to recreate the panel
-                        Ok(()) => self.preserve_state.switch_to_state(ctx, app),
-                        Err(err) => {
-                            self.error(ctx, app, format!("Couldn't save proposal: {}", err))
-                        }
-                    };
-                }
-                "Overwrite" => {
-                    // TODO If the user loaded the parent file again, this'll be confusing. Maybe
-                    // ban that?
-
-                    let proposals = &mut app.per_map.proposals;
-                    proposals.current_proposal.name =
-                        proposals.current_proposal.unsaved_parent.take().unwrap();
-
-                    return match inner_save(app) {
-                        Ok(()) => self.preserve_state.switch_to_state(ctx, app),
-                        // TODO If we fail to save for some reason, the Proposals panel gets out
-                        // of sync with the filesystem
-                        Err(err) => {
-                            self.error(ctx, app, format!("Couldn't save proposal: {}", err))
-                        }
-                    };
-                }
-                _ => unreachable!(),
-            },
-            Outcome::Changed(_) => {
-                self.name_updated(ctx);
-                Transition::Keep
-            }
-            _ => {
-                if ctx.normal_left_click() && ctx.canvas.get_cursor_in_screen_space().is_none() {
-                    return Transition::Pop;
-                }
-                Transition::Keep
-            }
-        }
-    }
-
-    fn draw_baselayer(&self) -> DrawBaselayer {
-        DrawBaselayer::PreviousState
-    }
-
-    fn draw(&self, g: &mut GfxCtx, app: &App) {
-        map_gui::tools::grey_out_map(g, app);
-        self.panel.draw(g);
-    }
-}
-
-fn inner_save(app: &App) -> Result<()> {
-    let proposal = &app.per_map.proposals.current_proposal;
-    let path = abstio::path_ltn_proposals(app.per_map.map.get_name(), &proposal.name);
-    let output_buffer = proposal.to_gzipped_bytes(app)?;
-    abstio::write_raw(path, &output_buffer)
-}
-
-fn load_picker_ui(
-    ctx: &mut EventCtx,
-    app: &App,
-    preserve_state: PreserveState,
-) -> Box<dyn State<App>> {
-    // Don't bother trying to filter out proposals currently loaded -- by loading twice, somebody
-    // effectively makes a copy to modify a bit
-    ChooseSomething::new_state(
-        ctx,
-        "Load which proposal?",
-        // basename (and thus list_all_objects) turn "foo.json.gz" into "foo.json", so further
-        // strip out the extension.
-        // TODO Fix basename, but make sure nothing downstream breaks
-        Choice::strings(
-            abstio::list_all_objects(abstio::path_all_ltn_proposals(app.per_map.map.get_name()))
-                .into_iter()
-                .map(abstutil::basename)
-                .collect(),
-        ),
-        Box::new(move |name, ctx, app| {
-            match Proposal::load_from_path(
-                ctx,
-                app,
-                abstio::path_ltn_proposals(app.per_map.map.get_name(), &name),
-            ) {
-                Some(err_state) => Transition::Replace(err_state),
-                None => preserve_state.switch_to_state(ctx, app),
-            }
-        }),
-    )
-}
-
 pub struct Proposals {
     // All entries are filled out, except for the current proposal being worked on
     list: Vec<Option<Proposal>>,
@@ -442,192 +227,6 @@ impl Proposals {
             }
         }
         false
-    }
-
-    pub fn to_widget_expanded(&self, ctx: &EventCtx, app: &App) -> Widget {
-        let mut col = Vec::new();
-        for (action, icon, hotkey) in [
-            ("New", "pencil", None),
-            ("Load (quick)", "folder", None),
-            ("Save (quick)", "save", Some(MultiKey::from(lctrl(Key::S)))),
-            ("Load (file)", "folder", None),
-            ("Save (file)", "save", None),
-            ("Share", "share", None),
-            ("Export GeoJSON", "export", None),
-        ] {
-            col.push(
-                ctx.style()
-                    .btn_plain
-                    .icon_text(&format!("system/assets/tools/{icon}.svg"), action)
-                    .hotkey(hotkey)
-                    .build_def(ctx),
-            );
-        }
-
-        for (idx, proposal) in self.list.iter().enumerate() {
-            let button = if let Some(proposal) = proposal {
-                ctx.style()
-                    .btn_solid_primary
-                    .text(format!("{} - {}", idx + 1, proposal.name))
-                    .hotkey(Key::NUM_KEYS[idx])
-                    .build_widget(ctx, &format!("switch to proposal {}", idx))
-            } else {
-                ctx.style()
-                    .btn_solid_primary
-                    .text(format!(
-                        "{} - {}",
-                        idx + 1,
-                        app.per_map.proposals.current_proposal.name
-                    ))
-                    .disabled(true)
-                    .build_def(ctx)
-            };
-            col.push(Widget::row(vec![
-                button,
-                // The first proposal (usually "existing LTNs", unless we're in a special consultation
-                // mode) is special and can't ever be removed
-                if idx != 0 {
-                    ctx.style()
-                        .btn_close()
-                        .disabled(self.list.len() == 1)
-                        .build_widget(ctx, &format!("hide proposal {}", idx))
-                } else {
-                    Widget::nothing()
-                },
-            ]));
-            // If somebody tries to load too many proposals, just stop
-            if idx == 9 {
-                break;
-            }
-        }
-        Widget::col(col)
-    }
-
-    pub fn to_widget_collapsed(&self, ctx: &EventCtx) -> Widget {
-        let mut col = Vec::new();
-        for (action, icon) in [
-            ("New", "pencil"),
-            ("Load (quick)", "folder"),
-            ("Save (quick)", "save"),
-            ("Share", "share"),
-            ("Export GeoJSON", "export"),
-        ] {
-            col.push(
-                ctx.style()
-                    .btn_plain
-                    .icon(&format!("system/assets/tools/{icon}.svg"))
-                    .build_widget(ctx, action),
-            );
-        }
-        Widget::col(col)
-    }
-
-    pub fn handle_action(
-        ctx: &mut EventCtx,
-        app: &mut App,
-        preserve_state: &PreserveState,
-        action: &str,
-    ) -> Option<Transition> {
-        match action {
-            "New" => {
-                // Fork a new proposal from the first one
-                if app.per_map.proposals.current != 0 {
-                    switch_to_existing_proposal(ctx, app, 0);
-                }
-
-                app.per_map.proposals.before_edit();
-            }
-            "Load (quick)" => {
-                return Some(Transition::Push(load_picker_ui(
-                    ctx,
-                    app,
-                    preserve_state.clone(),
-                )));
-            }
-            "Save (quick)" => {
-                return Some(Transition::Push(SaveDialog::new_state(
-                    ctx,
-                    app,
-                    preserve_state.clone(),
-                )));
-            }
-            "Load (file)" => {
-                let preserve_state = preserve_state.clone();
-                return Some(Transition::Push(FilePicker::new_state(
-                    ctx,
-                    Some(abstio::path_all_ltn_proposals(app.per_map.map.get_name())),
-                    Box::new(move |ctx, app, maybe_file| {
-                        match maybe_file {
-                            Ok(Some((path, bytes))) => {
-                                match Proposal::load_from_bytes(ctx, app, &path, Ok(bytes)) {
-                                    Some(err_state) => Transition::Replace(err_state),
-                                    None => preserve_state.switch_to_state(ctx, app),
-                                }
-                            }
-                            // No file chosen, just quit the picker
-                            Ok(None) => Transition::Pop,
-                            Err(err) => Transition::Replace(PopupMsg::new_state(
-                                ctx,
-                                "Error",
-                                vec![err.to_string()],
-                            )),
-                        }
-                    }),
-                )));
-            }
-            "Save (file)" => {
-                let proposal = &app.per_map.proposals.current_proposal;
-
-                return Some(Transition::Push(match proposal.to_gzipped_bytes(app) {
-                    Ok(contents) => FileSaver::with_default_messages(
-                        ctx,
-                        // * is used to indicate an unsaved file; don't include it in the filename
-                        format!("{}.json.gz", proposal.name.replace("*", "")),
-                        Some(abstio::path_all_ltn_proposals(app.per_map.map.get_name())),
-                        FileSaverContents::Bytes(contents),
-                    ),
-                    Err(err) => PopupMsg::new_state(ctx, "Save failed", vec![err.to_string()]),
-                }));
-            }
-            "Share" => {
-                return Some(Transition::Push(share::ShareProposal::new_state(ctx, app)));
-            }
-            "Export GeoJSON" => {
-                return Some(Transition::Push(match crate::export::geojson_string(app) {
-                    Ok(contents) => FileSaver::with_default_messages(
-                        ctx,
-                        format!("ltn_{}.geojson", app.per_map.map.get_name().map),
-                        None,
-                        FileSaverContents::String(contents),
-                    ),
-                    Err(err) => PopupMsg::new_state(ctx, "Export failed", vec![err.to_string()]),
-                }));
-            }
-            _ => {
-                if let Some(x) = action.strip_prefix("switch to proposal ") {
-                    let idx = x.parse::<usize>().unwrap();
-                    switch_to_existing_proposal(ctx, app, idx);
-                } else if let Some(x) = action.strip_prefix("hide proposal ") {
-                    let idx = x.parse::<usize>().unwrap();
-                    if idx == app.per_map.proposals.current {
-                        // First make sure we're not hiding the current proposal
-                        switch_to_existing_proposal(ctx, app, if idx == 0 { 1 } else { idx - 1 });
-                    }
-
-                    // Remove it
-                    app.per_map.proposals.list.remove(idx);
-
-                    // Fix up indices
-                    if idx < app.per_map.proposals.current {
-                        app.per_map.proposals.current -= 1;
-                    }
-                } else {
-                    return None;
-                }
-            }
-        }
-
-        Some(preserve_state.clone().switch_to_state(ctx, app))
     }
 }
